- Fix for #1165 - Reason for the issue: The freesolo Autocomplete component always returns a string when focus is lost. Another issue is that when the tab is switched, the string does not translated into the label. Open-Choice items were all modified to reflect this. We also explored on enabling autoSelect feaure on the Autocomplete. We discovered some other issues related to this so we decided not to use the autoSelect feature. https://github.com/mui/material-ui/issues/20602 - it highlights a selection as the user moves the mouse accidentally on an option. This can change options inadvertently.
So we don’t use the autoSelect feature. However, we need to have both onChange and onInputChange event listerener. In the InputChange event, the trick is to check for “reset” event and ignore the event as this is fired when the form is reset. Otherwises you will end up with looped call to setState.